Use binary release for bootstrap
At this point we have no other option: we need a working D compiler at
build time, but it seems like the older, pure-C++ LDC 0.17.x can't be
used to build LDC 1.30.0 and newer anymore.
And we can't recursively depend on DMD either, because when migrating to
a newer runtime we won't have DMD there.